Crate anonymiser

Anonymiser script

This is a script to be used to anonymise files on the local disk.

Install

There is an install.sh script which will download the latest binary from GitHub and add it to $HOME/.anonymiser/bin.

curl https://raw.githubusercontent.com/nationalarchives/dr2-court-document-package-anonymiser/main/install.sh | sh

You will need to add $HOME/.anonymiser/bin to your $PATH.

Running

anonymiser --input /path/to/input --output /path/to/output

Running with docker

docker run -v /path/to/input:/input -v /path/to/output:/output public.ecr.aws/u4s1g5v1/anonymiser

The input path must only contain the tar.gz files you’re converting.

Structs

Files 🔒

The input files and output directory

Functions

files_from_input_arguments 🔒

Process the input arguments

main 🔒

The entrypoint for the anonymiser script